refactor (ScreeningDate): ScreeningDate를 상영 날짜와 시간을 관리하는 클래스와 오브젝트로 구분
